pyproject.toml
[project]
name = "local-task-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg>=1.4",
    "openai-whisper",
    "scipy>=1.15.2",
    "sounddevice>=0.5.1",
]

[tool.uv.sources]
openai-whisper = { git = "https://github.com/openai/whisper.git" }
